refactor: utils/index 분리
- refactor: utils/index 각각 날짜, 숫자, 텍스트 util로 분리
