+++ src/pages/dashboard/hooks/useDashboardData.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 import { fetchDashboardData } from '@service/dashboard';
 import { useDemoStore } from '@stores/demoStore';
 
-import { formatYMD } from '@utils/index';
+import { formatYMD } from '@utils/dateUtils';
 import type { Register } from '@models/index';
 import type { TabOption } from './../type';
 import type { DashboardDataType } from '@service/dashboard/types';
