+++ src/pages/statistics/components/Emotion/EmotionalAmountChart.tsx
@@ -6,7 +6,7 @@ import StatisticsContentLayout from '../StatisticsContentLayout';
 import type { EmotionKey, Register } from '@models/index';
 import { Emotions, getEmotionText } from '@models/emotion';
 
-import { addCommasToNumber } from '@utils/index';
+import { addCommasToNumber } from '@utils/numberUtils';
 
 type EmotionalChartData = { type: EmotionKey; left: number; right: number }[];
 type EmotionChartProps = {
