+++ src/stores/services/demoService/util.ts
@@ -1,4 +1,5 @@
-import { compareYMDString, compareYMString, formatFromServer } from '@utils/index';
+import { compareYMDString, compareYMString, formatFromServer } from '@utils/dateUtils';
+
 import type { DemoExpenseDataType } from './types';
 import type { EmotionKey } from '@models/index';
 import type { ExpenseFilterType } from '@models/expense';
