+++ src/utils/dateUtils.ts
@@ -11,41 +11,6 @@ import {
   getMonth,
 } from 'date-fns';
 
-// #,##0 형식 formatter (숫자에 3자리씩 ',' 찍어주는 함수)
-export const addCommasToNumber = (number: number | string): string => {
-  if (number == undefined || number == null) return '0';
-  return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-};
-
-// #,##0 포맷된 string 또는 숫자를 다시 #,##0 으로 포맷
-export const formatAmountNumber = (
-  value: string,
-  hasLimit: boolean = false,
-  limit: number = 100000000,
-): string => {
-  /* case )
-   * 1. input의 handleChanged에서 #,##0 포맷된 string을 전달받음
-   * 2. 서버에서 넘어온 데이터는 숫자 형태의 string으로 넘어옴.
-   */
-  // 숫자만 포함된 값으로 변환
-  let cleanedValue = value.replace(/[^0-9]/g, '');
-
-  // 0으로 시작하지 않도록 제한
-  if (cleanedValue.startsWith('0')) {
-    cleanedValue = cleanedValue.replace(/^0+/, ''); // 0으로 시작할 경우 0->빈값으로 치환
-  }
-
-  // limit을 가지고 있다면, limit을 넘지 못하도록 제한 (제한 숫자로 바꿈)
-  if (hasLimit && cleanedValue.length > 0) {
-    const numericValue = parseInt(cleanedValue, 10); // 문자열을 정수로 변환
-    if (numericValue > limit) {
-      cleanedValue = limit.toString(); // 제한 값을 문자열로 변환
-    }
-  }
-
-  return addCommasToNumber(cleanedValue);
-};
-
 // Date 객체 -> yyyyMMdd String
 export const formatYMD = (date: Date, type: string = 'period'): string => {
   if (type === 'period') return format(date, 'yyyy.MM.dd');
@@ -92,9 +57,11 @@ export const compareYMString = (dateString: string, date: Date, format: string =
   return isSameYear(dateFromString, date) && getMonth(dateFromString) === getMonth(date);
 };
 
-export const convertToDateObject = (dateString: string, type: string = 'dash') => {
+// 'yyyy-MM-dd' 형태의 문자열을 date 객체로 돌려주는 함수
+export const convertToDateObject = (dateString: string, type: string = 'dash'): Date => {
   if (!dateString) return new Date();
   if (type === 'dash') return parse(dateString, 'yyyy-MM-dd', new Date());
+  return new Date();
 };
 
 // 서버에 보낼 때 날짜 형식 정의 `yyyy-MM-ddThh:mm:ss`
@@ -119,20 +86,4 @@ export const getTargetMonthDateObjArray = (date: Date) => {
 export const getThirtyDaysDateObjArray = (date: Date) => {
   const startDate = subDays(date, 29); // 전달받은 날짜 포함이라 29일을 빼서 시작
   return eachDayOfInterval({ start: startDate, end: date });
-};
-
-// 날짜, 내용, 금액, 지출 여부를 글로 돌려주는 함수 (spendDate = `yyyy-MM-ddThh:mm:ss` 형태)
-export const getSpendSumamryText = (
-  spendDate: string,
-  content: string,
-  amount: string,
-  regiesterType: string,
-) => {
-  return `${formatYMD(formatFromServer(spendDate))}, "${content}"에 ${addCommasToNumber(amount)}원 ${regiesterType === 'SPEND' ? '지출' : '절약'}`;
-};
-
-// 숫자 배열을 받으면, 각 항목이 전체에서 차지하는 비율 배열을 돌려준다.
-export const calculatePercentages = (data: number[]): number[] => {
-  const total = data.reduce((sum, value) => sum + value, 0);
-  return data.map((value) => parseFloat(((value / total) * 100).toFixed(1)));
 };
